discord_bot/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
MarkupSafe==2.1.5
multidict==6.0.5
py-cord==2.5.0
requests==2.31.0
typing_extensions==4.10.0
yarl==1.9.4
python-dotenv==1.0.1
